fix some StyleCop violations in common project - SA1503: Braces must not be omitted - SA1013: Closing brace must be preceded by a space. - SA1517: Code must not contain blank lines at start of file - SA1611: The documentation for parameter - SA1508: A closing brace must not be preceded by a blank line. - SA1128: Put constructor initializers on their own line - SA1642: Constructor summary documentation must begin with standard text - SA1208: Using directive for - SA1028: Code must not contain trailing whitespace - SA1515: Single-line comment must be preceded by blank line - SA1513: Closing brace must be followed by blank line - SA1309: Field must not begin with an underscore - SA1402: File may only contain a single type - SA1313: Parameter must begin with lower-case letter - SA1116: The parameters must begin on the line after the declaration, whenever the parameter span across multiple lines
