+++ src/GovITHub.Auth.Common/Data/ApplicationDbContext.cs
@@ -17,6 +17,7 @@ namespace GovITHub.Auth.Common.Data
         protected override void OnModelCreating(ModelBuilder builder)
         {
             base.OnModelCreating(builder);
+            
             // Customize the ASP.NET Identity model and override the defaults if needed.
             // For example, you can rename the ASP.NET Identity table names and more.
             // Add your customizations after calling base.OnModelCreating(builder);
