+++ src/GovITHub.Auth.Common/Data/ApplicationDbContextFactory.cs
@@ -1,10 +1,10 @@
-﻿using GovITHub.Auth.Common.Models;
+﻿using System.Reflection;
+using GovITHub.Auth.Common.Models;
 using IdentityServer4.EntityFramework.DbContexts;
 using IdentityServer4.EntityFramework.Options;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using MySQL.Data.Entity.Extensions;
-using System.Reflection;
 
 namespace GovITHub.Auth.Common.Data
 {
@@ -18,6 +18,7 @@ namespace GovITHub.Auth.Common.Data
             return new ApplicationDbContext(optionsBuilder.Options);
         }
     }
+
     public class ConfigurationDbContextFactory : IDbContextFactory<ConfigurationDbContext>
     {
         public ConfigurationDbContext Create(DbContextFactoryOptions options)
@@ -25,7 +26,6 @@ namespace GovITHub.Auth.Common.Data
             var optionsBuilder = new DbContextOptionsBuilder<ConfigurationDbContext>();
             var migrationsAssembly = typeof(ApplicationUser).GetTypeInfo().Assembly.GetName().Name;
             optionsBuilder.UseMySQL("DefaultConnection;", opts => opts.MigrationsAssembly(migrationsAssembly));
-
 
             return new ConfigurationDbContext(optionsBuilder.Options, new ConfigurationStoreOptions());
         }
