+++ src/GovITHub.Auth.Common/Data/Extensions/StringExtensions.cs
@@ -1,0 +1,21 @@
+﻿using System;
+
+namespace GovITHub.Auth.Common.Data
+{
+    public static class StringExtensions
+    {
+        public static string ToTitleCase(this string str)
+        {
+            if (string.IsNullOrWhiteSpace(str))
+                return str;
+            var tokens = str.Split(new[] { " " }, StringSplitOptions.RemoveEmptyEntries);
+            for (var i = 0; i < tokens.Length; i++)
+            {
+                var token = tokens[i];
+                tokens[i] = token.Substring(0, 1).ToUpper() + token.Substring(1);
+            }
+
+            return string.Join(" ", tokens);
+        }
+    }
+}
