+++ src/GovITHub.Auth.Common/Data/IConfigRepository.cs
@@ -1,6 +1,6 @@
-﻿using IdentityServer4.Models;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using IdentityServer4.Models;
 
 namespace GovITHub.Auth.Common.Data
 {
@@ -8,6 +8,5 @@ namespace GovITHub.Auth.Common.Data
     {
         Task<IEnumerable<Scope>> GetScopesAsync();
         Task<Client> GetClient(string id);
-
     }
 }
