+++ src/GovITHub.Auth.Common/Services/DeviceDetection/DeviceInfoBuilders/BrowserInfoBuilder.cs
@@ -6,7 +6,8 @@ namespace GovITHub.Auth.Common.Services.DeviceDetection.DeviceInfoBuilders
 {
     public class BrowserInfoBuilder : DeviceInfoBuilderBase<BrowserRegex>, IDeviceInfoBuilder
     {
-        public BrowserInfoBuilder(IDeviceInfoRegexLoader<BrowserRegex> regexLoader) : base(regexLoader)
+        public BrowserInfoBuilder(IDeviceInfoRegexLoader<BrowserRegex> regexLoader)
+            : base(regexLoader)
         {
         }
